Use Maven for building and modularize project
Maven is now the build tool of choice for the project.
This means that the project will use Maven to build and run the project instead of IntelliJ.

The project has also become modularized along with its previously non-modular dependencies.
By becoming a modular project, "jlink" is now able to create a custom runtime image.
The image can be wrapped as a .exe file for distribution.